requirements.txt
google-generativeai
python-dotenv
SpeechRecognition
pyttsx3
